feature: add product response dto to the project

+++ product-service/src/main/java/com/ecommerce/productservice/entity/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String productName;
 
     @Column(nullable = false)
-    private Long productPrice;
+    private Double productPrice;
 
     @Column(nullable = false)
     private Long productQuantity;
